requirements.txt
psycopg2
pydantic[email]
autopep8
SQLAlchemy-Utils
schedule
requests
passlib[bcrypt]
python-jose[cryptography]
python-multipart
mangum
pytest
uvicorn
urllib3
typing-extensions
toml
fastapi
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
h11==0.12.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
starlette==0.14.2
